python_libs/internet_speed/requirements.txt
pydantic==2.5.3
speedtest-cli==2.1.3
